size_t

+++ ft_printf/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: stomonoh <stomonoh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/16 17:58:01 by stomonoh          #+#    #+#             */
-/*   Updated: 2020/12/20 12:54:44 by stomonoh         ###   ########.fr       */
+/*   Updated: 2020/12/20 17:53:16 by stomonoh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <stdarg.h>
 # include <stdio.h>
 # include <unistd.h>
-# include "../../libft/libft.h"
+# include "../libft/libft.h"
 
 typedef struct{
 	int		flag;
@@ -29,8 +29,8 @@ int			ft_printf(const char *str, ...);
 void		print_type_c(t_in *test, unsigned char c, size_t *count);
 void		print_type_s(t_in *test, char *str, size_t *count);
 void		print_type_p(t_in *test, size_t p, size_t *count);
-void		print_type_x(t_in *test, unsigned int num, size_t *count);
-void		print_type_u(t_in *test, unsigned int num, size_t *count);
+void		print_type_x(t_in *test, size_t num, size_t *count);
+void		print_type_u(t_in *test, size_t num, size_t *count);
 void		print_type_id(t_in *test, int num, size_t *count);
 void		print_type_id_2(t_in *test, int len, int digit, long num, size_t *count);
 void		print_hex(char type, size_t num, size_t *count);
